agents/adk1003-team/pyproject.toml
[project]
name = "adk1003-team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[eval]>=0.2.0",
]
